+++ NEWS.md
@@ -3,6 +3,7 @@
 * Development version
 * Bug fixed in D_regularized_out
 * Addend.data argument added to all D_regularized -functions. In _fold -functions, test-partition of the data is appended, else the entire data frame is added.
+* Include ICC2 (group-mean reliability) for vpc_at. Enables calculation of sub-group mean-level reliabilities, in case the "at" is a group
 
 # multid 0.5.0